+++ worker/work.sh
@@ -1,0 +1,2 @@
+APP_URL=https://news.babataher.com WORKER_API_TOKEN=AAHUbqRiUer2CQpcef758gKm_4XaxjOkET8MAq1w2e311 ruby worker.rb
+